Flytt begrunnelser til felleskomponenter slik at den kan brukes for art16
