+++ src/felleskomponenter/begrunnelser.less
